terraform-serverless/service/requirements.txt
Flask==1.1.2
google-cloud-storage==1.35.0
gunicorn==20.0.4
requests==2.25.1